changed models.drill.ImageId to type int64

+++ models/drill.go
@@ -3,5 +3,5 @@ package models
 type Drill struct {
 	Title    string
 	Category string
-	ImageId  int
+	ImageId  int64
 }
